Changed name of FlightProgramInterpreter to FlightProgramParser and refactored it. Moved Command enum to its own file
