+++ Assets/Scripts/Interpreter/Command.cs
@@ -1,0 +1,17 @@
+namespace Interpreter
+{
+    public enum Command
+    {
+        Constant,
+        SetPos,
+        TakeOff,
+        FlyTo,
+        FlySpiral,
+        DroneMode,
+        FlyTrajectory,
+        Land,
+        Hover,
+        Eof,
+        Error
+    }
+}
